+++ app/helpers/kitten/components_helper.rb
@@ -3,7 +3,7 @@ module Kitten
     def render_group(group, title: nil)
       title ||= default_title(group)
       add_menu_group(group, title)
-      concat content_tag(:h2, class: 'karl-Title', id: group) { title }
+      concat content_tag(:h2, class: 'k-Row karl-Title', id: group) { title }
       render "kitten/groups/#{group}"
     end