+++ assets/javascripts/kitten/components/form/text-input-with-limit.js
@@ -10,21 +10,31 @@ export class TextInputWithLimit extends React.Component {
       value: this.props.defaultValue
     }
 
-    this.handleInput = this.handleInput.bind(this)
+    this.handleChange = this.handleChange.bind(this)
   }
 
-  handleInput(e) {
+  handleChange(e) {
     this.setState({ value: e.target.value })
+    this.props.onChange(e)
   }
 
   render() {
     const { limit,
             defaultValue,
+            onChange,
+            disabled,
             ...others } = this.props
 
     const length = this.state.value ? this.state.value.length : 0
 
-    const counterClassNames = classNames(
+    const textInputLimitClassName = classNames(
+      'k-TextInputLimit',
+      {
+        'is-disabled': disabled,
+      }
+    )
+
+    const counterClassName = classNames(
       'k-TextInputLimit__counter',
       {
         'is-error': length > limit,
@@ -32,12 +42,13 @@ export class TextInputWithLimit extends React.Component {
     )
 
     return (
-      <div className="k-TextInputLimit">
+      <div className={ textInputLimitClassName }>
         <TextInput className="k-TextInputLimit__input"
                    value={ this.state.value }
-                   onInput={ this.handleInput }
+                   onChange={ this.handleChange }
+                   disabled={ disabled }
                    { ...others } />
-        <div className={ counterClassNames }>
+        <div className={ counterClassName }>
           { limit - length }
         </div>
       </div>
@@ -46,6 +57,9 @@ export class TextInputWithLimit extends React.Component {
 }
 
 TextInputWithLimit.defaultProps = {
+  tag: 'input',
   limit: 80,
-  defaultValue: "",
+  defaultValue: '',
+  disabled: false,
+  onChange: function() {},
 }