+++ assets/javascripts/kitten/components/tooltips/tooltip.js
@@ -4,23 +4,36 @@ import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
 
 export default class Tooltip extends React.Component {
   render() {
+    const { place,
+            children,
+            id,
+            ...buttonTooltipIconProps } = this.props
+
     return (
       <div className="k-Tooltip">
         <ButtonTooltipIcon data-tip
-                           data-for={ this.props.id }
+                           data-for={ id }
                            data-event="click"
-                           data-dismiss={ this.props.id }
-                           aria-describedby={ this.props.id } />
-
-        <ReactTooltip id={ this.props.id }
+                           data-dismiss={ id }
+                           aria-describedby={ id }
+                           { ...buttonTooltipIconProps } />
+        <ReactTooltip id={ id }
+                      // This is not a mistake, this attribute is called
+                      // class not className!
                       class="k-Tooltip__content"
                       role="tooltip"
                       effect="solid"
-                      place={ this.props.place }
-                      globalEventOff="click">
-          { this.props.children }
+                      place={ place }
+                      event="none">
+          { children }
         </ReactTooltip>
       </div>
     )
   }
+}
+
+Tooltip.defaultProps = {
+  id: '',
+  place: 'right',
+  children: null,
 }