+++ assets/javascripts/kitten/components/typography/paragraph.js
@@ -3,19 +3,29 @@ import classNames from 'classnames'
 
 export class Paragraph extends React.Component {
   render() {
-    const { className, tag, modifier, ...other } = this.props
+    const {
+      className,
+      tag,
+      modifier,
+      margin,
+      normalLineHeight,
+      ...other,
+    } = this.props
 
     const paragraphClassNames = classNames(
       'k-Paragraph',
       className,
       `k-Paragraph--${modifier}`,
+      {
+        'k-Paragraph--withoutMargin': !margin,
+        'k-Paragraph--normalLineHeight': normalLineHeight,
+      }
     )
 
     const Tag = tag
 
     return (
-      <Tag className={ paragraphClassNames }
-        { ...other } />
+      <Tag className={ paragraphClassNames } { ...other } />
     )
   }
 }
@@ -23,4 +33,6 @@ export class Paragraph extends React.Component {
 Paragraph.defaultProps = {
   tag: 'p',
   modifier: 'primary',
+  margin: true,
+  normalLineHeight: false,
 }